PWN-3245 - Add history tab to bottom navigation and rename files (#415)
