+++ app/src/main/java/org/p2p/wallet/history/ui/token/adapter/HistoryAdapterPagingController.kt
@@ -1,4 +1,4 @@
-package org.p2p.wallet.history.ui.history.adapter
+package org.p2p.wallet.history.ui.token.adapter
 
 import org.p2p.wallet.common.ui.recycler.PagingState
