+++ app/src/main/java/org/p2p/wallet/history/ui/token/adapter/holders/DateViewHolder.kt
@@ -1,4 +1,4 @@
-package org.p2p.wallet.history.ui.history.adapter.holders
+package org.p2p.wallet.history.ui.token.adapter.holders
 
 import android.view.ViewGroup
 import org.p2p.wallet.common.date.toDateString
