+++ app/src/main/java/org/p2p/wallet/history/ui/token/adapter/holders/HistoryTransactionViewHolder.kt
@@ -1,4 +1,4 @@
-package org.p2p.wallet.history.ui.history.adapter.holders
+package org.p2p.wallet.history.ui.token.adapter.holders
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
