+++ app/src/main/res/drawable/ic_home_history.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="25dp"
+    android:height="24dp"
+    android:viewportWidth="25"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="@color/textIconSecondary"
+        android:fillType="evenOdd"
+        android:pathData="M17.6387 5.86865C16.1533 4.62373 14.2665 3.96037 12.3288 4.00183C10.3912 4.0433 8.53444 4.78678 7.10366 6.09412C5.67288 7.40145 4.76534 9.18375 4.54968 11.1098C4.48823 11.6587 3.99347 12.0538 3.44462 11.9923C2.89576 11.9309 2.50065 11.4361 2.5621 10.8873C2.83167 8.47968 3.9661 6.25182 5.75457 4.61764C7.54305 2.98347 9.86396 2.05413 12.286 2.00229C14.7081 1.95046 17.0667 2.77966 18.9234 4.33582C20.7802 5.89198 22.0089 8.06927 22.3812 10.4631C22.7535 12.857 22.2442 15.3046 20.9478 17.3512C19.6514 19.3978 17.6561 20.9042 15.3327 21.5904C13.0093 22.2767 10.5157 22.0961 8.31542 21.0824C6.79692 20.3827 5.48934 19.3191 4.5 18V20C4.5 20.5523 4.05228 21 3.5 21C2.94772 21 2.5 20.5523 2.5 20V15C2.5 14.4477 2.94772 14 3.5 14H8.5C9.05228 14 9.5 14.4477 9.5 15C9.5 15.5523 9.05228 16 8.5 16H5.57178C6.39549 17.4268 7.64382 18.5709 9.15234 19.2659C10.9126 20.0769 12.9074 20.2213 14.7661 19.6723C16.6249 19.1233 18.2211 17.9182 19.2582 16.281C20.2953 14.6437 20.7028 12.6856 20.405 10.7705C20.1071 8.85542 19.1241 7.11358 17.6387 5.86865ZM13.5 8C13.5 7.44772 13.0523 7 12.5 7C11.9477 7 11.5 7.44772 11.5 8V12.9993V13L11.5 13.003C11.5008 13.2769 11.6117 13.5248 11.7908 13.705L13.5 8ZM13.5 8V12.5858L15.7126 14.7984C16.1032 15.1889 16.1032 15.8221 15.7126 16.2126C15.3221 16.6032 14.6889 16.6032 14.2984 16.2126L11.7934 13.7076L13.5 8Z" />
+</vector>
