+++ app/src/main/res/menu/main.xml
@@ -4,6 +4,11 @@
         android:id="@+id/itemHome"
         android:icon="@drawable/ic_home_main"
         android:title="@string/home_main" />
+
+    <item
+        android:id="@+id/itemHistory"
+        android:icon="@drawable/ic_home_history"
+        android:title="@string/home_history"/>
     <item
         android:id="@+id/itemSend"
         android:icon="@drawable/ic_home_send"
